# backend/gateway_service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Gateway банковского приложения. Единая точка входа для клиентских запросов.",
        "title": "Gateway Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/api/v1/auth/pins": {
            "put": {
                "security": [
                    {
                        "SessionToken": []
                    }
                ],
                "description": "Устанавливает новый или изменяет существующий 4-значный PIN-код.\nОперация доступна только аутентифицированным пользователям.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Обновить PIN-код",
                "parameters": [
                    {
                        "description": "Новый 4-значный PIN",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.SetPinRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PIN успешно изменен",
                        "schema": {
                            "$ref": "#/definitions/schemas.SuccessResponse"
                        }
                    },
                    "401": {
                        "description": "Сессия невалидна или отсутствует токен",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/self-block": {
            "post": {
                "security": [
                    {
                        "SessionToken": []
                    }
                ],
                "description": "Полностью блокирует доступ к аккаунту и замораживает все счета.\nВНИМАНИЕ: Для разблокировки потребуется обращение в поддержку или использование email-кода.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Самоблокировка клиента",
                "responses": {
                    "200": {
                        "description": "Аккаунт успешно заблокирован",
                        "schema": {
                            "$ref": "#/definitions/schemas.SuccessResponse"
                        }
                    },
                    "401": {
                        "description": "Необходима авторизация",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/unlock-codes": {
            "post": {
                "description": "Отправляет 6-значный одноразовый код на привязанный Email пользователя.\nКод необходим для сброса блокировки после неверного ввода PIN.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth-unlock"
                ],
                "summary": "Запросить код разблокировки",
                "parameters": [
                    {
                        "description": "Email, привязанный к аккаунту",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.RequestUnlockRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Код успешно отправлен",
                        "schema": {
                            "$ref": "#/definitions/schemas.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный Email",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Пользователь с таким Email не найден",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/unlock-codes/verifications": {
            "post": {
                "description": "Проверяет код из письма и разблокирует учетную запись пользователя.\nПосле разблокировки пользователь сможет снова войти по своему PIN.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth-unlock"
                ],
                "summary": "Подтвердить разблокировку",
                "parameters": [
                    {
                        "description": "Email и 6-значный код",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.UnlockRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Аккаунт успешно разблокирован",
                        "schema": {
                            "$ref": "#/definitions/schemas.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный или просроченный код",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/sessions": {
            "post": {
                "description": "Аутентификация по номеру телефона и PIN-коду.\nВ случае успеха возвращает сессионный токен, который нужно передавать в заголовке X-Session-Token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sessions"
                ],
                "summary": "Вход в приложение",
                "parameters": [
                    {
                        "description": "Номер телефона (+7...) и 4-значный PIN",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.LoginPinRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Успешная авторизация",
                        "schema": {
                            "$ref": "#/definitions/schemas.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный формат входных данных",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Неверный телефон или PIN-код",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "423": {
                        "description": "Аккаунт временно заблокирован из-за перебора PIN",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "SessionToken": []
                    }
                ],
                "description": "Аннулирует ВСЕ активные сессии текущего пользователя.\nПолезно в случае компрометации одного из устройств.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sessions"
                ],
                "summary": "Выход со всех устройств",
                "responses": {
                    "200": {
                        "description": "Все сессии завершены",
                        "schema": {
                            "$ref": "#/definitions/schemas.SuccessResponse"
                        }
                    },
                    "401": {
                        "description": "Необходима авторизация",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/sessions/current": {
            "delete": {
                "security": [
                    {
                        "SessionToken": []
                    }
                ],
                "description": "Удаляет текущий сессионный токен из хранилища Redis.\nПосле этого токен станет недействительным для всех последующих запросов.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sessions"
                ],
                "summary": "Выход (текущая сессия)",
                "responses": {
                    "200": {
                        "description": "Выход успешно выполнен",
                        "schema": {
                            "$ref": "#/definitions/schemas.SuccessResponse"
                        }
                    },
                    "401": {
                        "description": "Токен уже недействителен",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "schemas.ErrorResponse": {
            "type": "object",
            "properties": {
                "detail": {
                    "description": "Подробное сообщение об ошибке для разработчика и пользователя",
                    "type": "string",
                    "example": "Неверный PIN-код или формат данных"
                }
            }
        },
...
        "schemas.LoginPinRequest": {
            "type": "object",
            "required": [
                "phone",
                "pin"
            ],
            "properties": {
                "phone": {
                    "description": "Номер телефона в международном формате (+7...)",
                    "type": "string",
                    "format": "phone",
                    "example": "+79991234567"
                },
                "pin": {
                    "description": "Секретный 4-значный код доступа",
                    "type": "string",
                    "maxLength": 4,
                    "minLength": 4,
                    "example": "1234"
                }
            }
        },
        "schemas.LoginResponse": {
            "type": "object",
            "properties": {
                "has_pin": {
                    "description": "Признак наличия установленного PIN-кода (false - если нужно установить впервые)",
                    "type": "boolean",
                    "example": true
                },
                "session_token": {
                    "description": "Сессионный токен (JWT), необходим для авторизации последующих запросов",
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                },
                "status": {
                    "description": "Текущий статус аккаунта в системе",
                    "type": "string",
                    "example": "active"
                }
            }
        },
...
        "schemas.RequestUnlockRequest": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "description": "Email, привязанный к профилю клиента",
                    "type": "string",
                    "format": "email",
                    "example": "user@example.com"
                }
            }
        },
        "schemas.SetPinRequest": {
            "type": "object",
            "required": [
                "pin"
            ],
            "properties": {
                "pin": {
                    "description": "Новый 4-значный секретный код",
                    "type": "string",
                    "maxLength": 4,
                    "minLength": 4,
                    "example": "1234"
                }
            }
        },
        "schemas.SuccessResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "description": "Статус успеха (обычно 'success')",
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "schemas.UnlockRequest": {
            "type": "object",
            "required": [
                "code",
                "email"
            ],
            "properties": {
                "code": {
                    "description": "6-значный код подтверждения из Email-письма",
                    "type": "string",
                    "maxLength": 6,
                    "minLength": 6,
                    "example": "123456"
                },
                "email": {
                    "description": "Email, привязанный к профилю клиента",
                    "type": "string",
                    "format": "email",
                    "example": "user@example.com"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "OnboardingToken": {
            "description": "Токен регистрации, полученный на Шаге 0 (Start Onboarding)",
            "type": "apiKey",
            "name": "X-Onboarding-Token",
            "in": "header"
        },
        "SessionToken": {
            "description": "Сессионный токен, полученный при авторизации",
            "type": "apiKey",
            "name": "X-Session-Token",
            "in": "header"
        }
    },
...
}